fix(wsl2): resolve PID translation bug and fix seq type mismatch in tests
- Add BPF source code (bpf/loader/bpfsrc/pid_check.bpf.c) with build
  instructions comment in wsl2_loader.go for provenance tracking
- Fix seq type mismatch in AssertKernEvent: cast both sides to uint64
  to prevent false failures comparing uint32 vs uint64
- Update HANDOVER.md: document WSL2 PID translation fix (§20),
  update known limitations (§13.3), and record test results (23/25 pass)

+++ agent/agent_utils_test.go
@@ -276,7 +276,7 @@ func AssertKernEvent(t *testing.T, kernEvt *bpf.AgentKernEvt, conditions KernDat
 	}
 	seq := kernEvt.Seq
 	if !conditions.ignoreSeq {
-		assert.Equal(t, conditions.seq, seq)
+		assert.Equal(t, uint64(conditions.seq), uint64(seq))
 	}
 	step := kernEvt.Step
 	if !conditions.ignoreStep {
